+++ migrations/20220605170157_create_create_orders_tables.postgres.up.sql
@@ -3,7 +3,7 @@ CREATE TABLE orders
     id          SERIAL PRIMARY KEY,
     number      VARCHAR UNIQUE,
     status      VARCHAR   NOT NULL,
-    accrual     INT,
+    accrual     NUMERIC(32, 2),
     uploaded_at TIMESTAMPTZ NOT NULL,
     user_id     SERIAL    NOT NULL,
     CONSTRAINT fk_user
